Refactored src to create viewers with factory function pattern. Core viewer logic is now completely reusable with all kinds of model formats.
