+++ demo/src/App.js
@@ -5,12 +5,28 @@ import Highlighting from './component-demos/Highlighting';
 import NavCube from './component-demos/NavCube';
 import Screenshot from './component-demos/Screenshot';
 import ConditionalRender from './component-demos/ConditionalRender';
+import GLTFModel from './component-demos/GLTFModel';
+import XKTModel from './component-demos/XKTModel';
 
 const App = () => (
   <div className="container my-5">
     <h1 className="text-center text-light my-5">
       Xeokit-SDK React integration demo
     </h1>
+    <div className="card bg-light my-5">
+      <div className="card-header">Feature</div>
+      <div className="card-body">
+        <h4 className="card-title mb-5">GLTF Model</h4>
+        <GLTFModel />
+      </div>
+    </div>
+    <div className="card bg-light my-5">
+      <div className="card-header">Feature</div>
+      <div className="card-body">
+        <h4 className="card-title mb-5">XKT Model</h4>
+        <XKTModel />
+      </div>
+    </div>
     <div className="card bg-light my-5">
       <div className="card-header">Feature</div>
       <div className="card-body">
@@ -43,7 +59,7 @@ const App = () => (
       <div className="card-header">Feature</div>
       <div className="card-body">
         <h4 className="card-title mb-5">Take screenshot of scene</h4>
-        <Screenshot />
+        {/* <Screenshot /> */}
       </div>
     </div>
     <div className="card bg-light my-5">
