+++ demo/src/models.js
@@ -1,20 +1,28 @@
 export const sampleModel1 = {
-    id: "duplex",
-    src: "./models/duplex/scene.gltf",
-    metaModelSrc: "./metaModels/duplex/metaModel.json",
-    edges: true
+  id: 'duplex',
+  src: './models/duplex/scene.gltf',
+  metaModelSrc: './metaModels/duplex/metaModel.json',
+  edges: true,
 };
 
 export const sampleModel2 = {
-    id: "OTCConferenceCenter",
-    src: "./models/OTCConferenceCenter/scene.gltf",
-    metaModelSrc: "./metaModels/OTCConferenceCenter/metaModel.json",
-    edges: true
+  id: 'OTCConferenceCenter',
+  src: './models/OTCConferenceCenter/scene.gltf',
+  metaModelSrc: './metaModels/OTCConferenceCenter/metaModel.json',
+  edges: true,
 };
 
 export const sampleModel3 = {
-    id: "schependomlaan",
-    src: "./models/schependomlaan/scene.gltf",
-    metaModelSrc: "./metaModels/schependomlaan/metaModel.json",
-    edges: true
+  id: 'schependomlaan',
+  src: './models/schependomlaan/scene.gltf',
+  metaModelSrc: './metaModels/schependomlaan/metaModel.json',
+  edges: true,
+};
+
+export const xktModel = {
+  id: 'xkt-model',
+  src: './xkt/duplex.xkt',
+  metaModelSrc: './xkt/duplex.json',
+  edges: true,
+  performacne: true,
 };
